MAINT: Return `NaN` or `Inf` for out of range parameter values when `disable_checks=True`.
Currently, the behavior of `random_polyagamma` when called with the
`disable_checks=True` flag is "anything goes"; meaning that it might
return a 0.0, crash your program or hang forever if the shape parameter
is not positive. This is outlined in the docstring:

```
    disable_checks : bool, optional
        Whether to check that the `h` parameter contains only positive
        values(s). Disabling may give a performance gain, but may result
        in problems (crashes, non-termination, wrong return values)
        if the inputs do contain non-positive values.
```

 I noticed that the behavior typically depends on the algorithm being used:
 - `devroye` and `gamma`: It always returns 0 if the  `h` parameter is out of range.
 - `saddle` and `alternate`: It always hangs forever and requires one to
 kill the interpreter manually. This is because the accept-rejection step
 in the algorithm never terminates due to the out-of-range value of `h`.

This commit ensures that under the above conditions, an explicit `NaN`
or `Inf` is returned to indicate the invalidity of the input parameter.