BUILD,CI: Use latest numpy version to build wheels.
This ensures we use the latest available numpy version to build wheels.
We also no longer build the wheels on ARM64 arch of MacOS since they are built
on the x86_64 arch and thus cannot be tested on the CI release pipeline.
As recommended by the CIBUILDWHEEL team, we have users build the ARM64
wheels on MacOS natively during installation of the package via source
distribution. Both the above address the bug reported in
https://github.com/zoj613/polyagamma/issues/129, since the bug
originates from numpy's 2.1.3 version used to build wheels of version
2.0.1.